\docType{data}
\name{known.encodings}
\alias{.known.encodings}
\alias{known.encodings}
\title{Known gene/protein encodings, i.e. type of identifiers such as Ensembl or Entrez.}
\format{chr [1:3] "ensembl" "entrez" "refseq"}
\usage{
  .known.encodings
}
\description{
  Known gene/protein encodings, i.e. type of identifiers
  such as Ensembl or Entrez.
}
\keyword{datasets}

