\docType{data}
\name{known.encodings}
\alias{known.encodings}
\title{Known gene/protein encodings/identifiers}
\format{chr [1:3] "ensembl" "entrez" "refseq"}
\usage{
  known.encodings
}
\description{
  In these packages, we call the type of identifier for an
  biological entity `encodings'. E.g. Entrez, Ensembl or
  Refseq are examples of encodings.
}
\keyword{datasets}

